+++ _Typings/gameApi/alias/utils.d.ts
@@ -2,9 +2,9 @@
 // For details: https://github.com/Withered-Flower-0422/BST/blob/main/LICENSE
 
 declare module "game:alias" {
-    export { KeyboardKey, MouseButton, GamePadButton, PlayerKey } from "keys"
+    export { KeyboardKey, NonNoneKey, MouseButton, GamePadButton, PlayerKey } from "keys"
 
     export { CancelableEvent, VfxType } from "game:module"
 
-    export { BallType, CollisionEvent, EaseType, Events, Item, RaycastResult, Tween } from "game:type"
+    export { BallType, CollisionEvent, EaseType, Events, RegisterEvents, Item, RaycastResult, Tween } from "game:type"
 }